Makefile: add an automated pipeline for generation
If we are ever to update the source OpenAPI specification used to
generate the models and client in this repository, we need an automated
pipeline to do so. The Makefile targets in this repository automate
every step of that, from downloading the source manifests to converting
them to v3 and generating Go code from them.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>